Don't allow external API usage
