+++ utils/gear-tasks/Cargo.toml
@@ -23,7 +23,10 @@ futures-executor = { version = "0.3.30", default-features = false, optional = tr
 log.workspace = true
 
 [dev-dependencies]
-sp-io.workspace = true
+gear-tasks = { workspace = true, features = ["std"] }
+
+sp-io = { workspace = true, features = ["std"] }
+sp-state-machine = { workspace = true, features = ["std"] }
 futures-executor = "0.3.30"
 gear-node-testing.workspace = true
 substrate-test-client.workspace = true
@@ -39,6 +42,7 @@ std = [
     "sp-runtime-interface/std",
     "sp-externalities/std",
     "sp-api/std",
+    "gear-tasks-runtime-api/std",
     "frame-support/std",
     "parity-scale-codec/std",
     "futures-executor",
