+++ utils/gear-tasks/runtime-api/src/lib.rs
@@ -27,19 +27,3 @@ sp_api::decl_runtime_apis! {
         fn execute_task(func_ref: u64, payload: Vec<u8>) -> Vec<u8>;
     }
 }
-
-pub fn impl_fn(func_ref: u64, payload: Vec<u8>) -> Vec<u8> {
-    #[cfg(target_arch = "wasm32")]
-    let f = unsafe { core::mem::transmute::<u32, fn(Vec<u8>) -> Vec<u8>>(func_ref as u32) };
-
-    // used only in tests
-    #[cfg(feature = "std")]
-    let f = unsafe { core::mem::transmute::<u64, fn(Vec<u8>) -> Vec<u8>>(func_ref) };
-
-    // tasks must only read storage and never write
-    let output_payload;
-    frame_support::assert_storage_noop!({
-        output_payload = f(payload);
-    });
-    output_payload
-}
